Install webpack + npm to begin the front end.
Place views into the client/views folder. Each view should consist of a
folder of the name of the view, an index.js file containing the entry
point of the view, and all needed assets. Be sure to require() all
needed assets in the entry point---yes, even static files like
.css/.less/etc. and image formats like .png and .jpg. HTML file with
the name of the view will be auto-generated with references to required
assets replaced with the production, hash-named versions. Currently
this file is generated with a default template with no content in the
<body> tag. In the future the system will look for a file named index
(not index.js) and will load the template with the appropriate template
parser. For now, add a <div id='root'></div> node using the entry point
file and attach all shadow dom to that node. This can be accompilshed
with the following sample code:

    // index.js
    import { render } from 'react-dom';
    import Component from './component';

    var root;
    (root = document.createElement('div')).setAttribute('id', 'root');
    document.body.appendChild(root);

    render(<Component />, root)

Furthermore, change the title of the page programmatically by modifying
document.title.

A note on webpack:

Webpack, according to their website, is a tool which "takes modules
with dependencies and generates static assets representing those
modules." Concretely, webpack allows us to specify .js files, called
entry points, which define dependencies as well as run code on the
website. A plugin called html-webpack-plugin generates HTML containing
references to these dependencies.

Webpack also does other useful things, such as bundles and exports
files, as well as adds hashes to the file names of static assets in
order to better support long term hashing. At a high level, it
"complies" each view.

The way webpack functions is defined by the webpack.config.js file
located in the root of the directory. Open the file for futher
information on how webpack runs.