Refactor logged in user authentication
Previously users viewed a Flash of Logged Out Content every time they
navigated to a new page. The system for managing the user context (as
well as indirectly the elements which show up on every page; i.e., the
header and the footer) have been refactored out into a seperate
component called UserContext in Routes.js. It passes down the user prop
to all its children as well as a logout callback function in order that
these flashes no longer appear as well as unnecessary HTTP request do
not get made.

Some small changes include a unsaved notification in Dashboard if any
teams are currently open, a clarification dialog on the "combine"
checkbox, as well as info on where to send payment under the "confirm
registration" button.